Add a number of likes bar and liking/unliking functionality to the like button
Added a likes bar that would display if the the post has 1 or more likes it. In addition, a liking and unliking functionality has been added to the thumbs up icon. The Thumbs Up icon and the "Like" word will be highlighted in blue if the user visiting the page has liked the post.